Upgrade to express 4, move to templates
In order to do the upgrade to express 4 for built apis, additional
dependencies are required since middleware is no longer bundled with
express. Trying to manage those dependencies through the existing inline
file construction behavior would have been really cumbersome, so the
file contents for built apis are now based on templates.

A nice side effect of this change is that mockapi itself no longer needs
to explicitly depend on express.